libs: make every implicit integer conversion explicit
Preparation for promoting -Wconversion to a hard error (#240). Five
mechanical classes, all semantics-preserving:

- register accessors added ra8_*_addr_t + ra8_*_off_t enums directly;
  cast both operands to uintptr_t (clang -Wenum-enum-conversion), same
  for the other mixed-enum arithmetic in ra8_rar / ra8_rsip / ra8_usb_hhid
- 1UL/0x3UL shift-mask idioms stored into uint32_t are 64-bit on the
  unit-test host (clang/gcc -Wshorten-64-to-32 / -Wconversion); use the
  32-bit 1U/0x3U literals the target semantics always assumed
- 0U/1U ternaries assigned to uint8_t promote to int; cast the result
- ra8_reflow_internal_xhtml_invalid took uint32_t while its only caller
  passes the size_t engine->xhtml_len; widen the parameter to size_t
  (and include stddef.h in the internal header)
- complete the ra8_fs_backend_t initializer in exfat_fs_test (erase_blocks
  hook was silently dropped by the 4-element list)

+++ libs/ra8_core/src/ra8_pin_validator.c
@@ -35,7 +35,8 @@ static const char* s_tag = "PINVAL";
  * @note Size is exactly `k_ra8_port_count * k_ra8_pin_count /
  *       k_ra8_bits_per_byte` bytes = 30.
  */
-static uint8_t s_claimed[(k_ra8_port_count * k_ra8_pin_count) / k_ra8_bits_per_byte];
+static uint8_t s_claimed[((uint16_t)k_ra8_port_count * (uint16_t)k_ra8_pin_count) /
+                         (uint16_t)k_ra8_bits_per_byte];
 
 /**
  * @var s_owner
@@ -45,7 +46,7 @@ static uint8_t s_claimed[(k_ra8_port_count * k_ra8_pin_count) / k_ra8_bits_per_b
  * `NULL` means "not claimed". Claim stores the caller-provided string
  * literal pointer; the validator never copies the bytes.
  */
-static const char* s_owner[k_ra8_port_count * k_ra8_pin_count];
+static const char* s_owner[(uint16_t)k_ra8_port_count * (uint16_t)k_ra8_pin_count];
 
 /* Convert a packed `ra8_port_pin_t` to a flat bit index -- see implementation for details. */
 static ra8_err_t internal_flat_index(ra8_port_pin_t pin, uint16_t* out_index)
@@ -201,7 +202,7 @@ void ra8_pin_validator_reset(void)
   for (uint16_t i = 0U; i < (uint16_t)sizeof(s_claimed); i++) {
     s_claimed[i] = 0U;
   }
-  for (uint16_t i = 0U; i < (uint16_t)(k_ra8_port_count * k_ra8_pin_count); i++) {
+  for (uint16_t i = 0U; i < ((uint16_t)k_ra8_port_count * (uint16_t)k_ra8_pin_count); i++) {
     s_owner[i] = nullptr;
   }
 }
